fix: restore dashboard auth and suppress PersistedQueryNotFound spam
