+++ .env.example
@@ -23,6 +23,11 @@ PORT=9090
 # DASHBOARD_USER=admin
 # DASHBOARD_PASSWORD_SHA256=5e884898da28047151d0e56f8dc6292773603d0d6aabbdd62a11ef721d1542d8
 
+# API key for machine-to-machine auth (dashboard ↔ bot).
+# Accepted via X-API-Key header as an alternative to Basic Auth.
+# Generate with: node -e "console.log(require('crypto').randomBytes(32).toString('hex'))"
+# DASHBOARD_API_KEY=
+
 # Global notification secrets (shared by all accounts unless overridden per-account)
 TELEGRAM_TOKEN=123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11
 TELEGRAM_CHAT_ID=987654321
